add(registry): all the themes...
for themes whose keys conflict, a good example is gruvbox or nord, I
tried gave the keys a unique name which express what they theme they
are, without deviating too much from the repo name.

Also, the pkgname field and the key of the scheme they correspond to
will be the same. I will write a test for this.